pyproject.toml
[tool.poetry]
name = "ice-breaker"
version = "0.1.0"
description = ""
authors = ["Darlington02 <darlingtonnnam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
langchain-openai = "^0.2.5"
langchain-community = "^0.3.5"
langchainhub = "^0.1.21"
python-dotenv = "^1.0.1"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
